Move code to more granular locations.
