+++ test/test_decompose_loop.py
@@ -1,4 +1,3 @@
-from typing import cast
 import numpy as np
 from unittest.mock import Mock, patch
 from pytest import raises, LogCaptureFixture
@@ -7,13 +6,10 @@ from fi_nomad.kernels import KernelBase
 from fi_nomad.entry import (
     compute_max_iterations,
     decompose,
-    initialize_candidate,
-    instantiate_kernel,
 )
 
 from fi_nomad.types import (
     BaseModelFreeKernelReturnType,
-    FloatArrayType,
     InitializationStrategy,
     KernelInputType,
     KernelReturnType,
@@ -48,116 +44,9 @@ class TestKernel(KernelBase):
         )
 
 
-## Kernel Instantiation
-def test_instantiate_kernel_throws_on_unsupported_kernel() -> None:
-    mock_data_in = Mock()
-    with raises(ValueError, match="Unsupported"):
-        _ = instantiate_kernel(KernelStrategy.TEST, mock_data_in)
-
-
-def test_instantiate_kernel_throws_on_unsupported_feature() -> None:
-    mock_data_in = Mock()
-    some_non_null_value = 0.5
-    with raises(NotImplementedError):
-        _ = instantiate_kernel(
-            KernelStrategy.BASE_MODEL_FREE, mock_data_in, some_non_null_value
-        )
-
-
-@patch(f"{PKG}.RowwiseVarianceGaussianModelKernel", autospec=True)
-@patch(f"{PKG}.SingleVarianceGaussianModelKernel", autospec=True)
-@patch(f"{PKG}.BaseModelFree", autospec=True)
-def test_instantiate_kernel_honors_strategy_selection(
-    mock_base: Mock, mock_svgauss: Mock, mock_rwgauss: Mock
-) -> None:
-    mock_data_in = Mock()
-    _ = instantiate_kernel(KernelStrategy.BASE_MODEL_FREE, mock_data_in)
-    mock_base.assert_called_once()
-    _ = instantiate_kernel(KernelStrategy.GAUSSIAN_MODEL_SINGLE_VARIANCE, mock_data_in)
-    mock_svgauss.assert_called_once()
-    _ = instantiate_kernel(KernelStrategy.GAUSSIAN_MODEL_ROWWISE_VARIANCE, mock_data_in)
-    mock_rwgauss.assert_called_once()
-
-
-## low-rank candidate initialization
-@patch(f"{PKG}.initialize_low_rank_candidate")
-def test_initialize_candidate_honors_initialization_strategy(mock_init: Mock) -> None:
-    mock_x = np.array([[0, 2], [2, 0]])
-    mock_checkpoint = np.eye(2)
-    with raises(AssertionError):
-        np.testing.assert_array_equal(mock_x, mock_checkpoint)
-    k_strat = KernelStrategy.GAUSSIAN_MODEL_SINGLE_VARIANCE
-
-    _ = initialize_candidate(
-        InitializationStrategy.COPY, k_strat, mock_x, mock_checkpoint
-    )
-    mock_init.assert_called_once_with(mock_x, InitializationStrategy.COPY)
-    mock_init.reset_mock()
-
-    _ = initialize_candidate(
-        InitializationStrategy.BROADCAST_MEAN, k_strat, mock_x, mock_checkpoint
-    )
-    mock_init.assert_called_once_with(mock_x, InitializationStrategy.BROADCAST_MEAN)
-    mock_init.reset_mock()
-
-    _ = initialize_candidate(
-        InitializationStrategy.KNOWN_MATRIX, k_strat, mock_x, mock_checkpoint
-    )
-    mock_init.assert_called_once_with(
-        mock_checkpoint, InitializationStrategy.KNOWN_MATRIX
-    )
-
-
-@patch(f"{PKG}.initialize_low_rank_candidate")
-def test_initialize_candidate_forces_copy_strategy_for_base_model_free(
-    mock_init: Mock,
-) -> None:
-    mock_x = np.array([[0, 2], [2, 0]])
-    mock_checkpoint = Mock()
-    k_strat = KernelStrategy.BASE_MODEL_FREE
-
-    _ = initialize_candidate(
-        InitializationStrategy.COPY, k_strat, mock_x, mock_checkpoint
-    )
-    mock_init.assert_called_once_with(mock_x, InitializationStrategy.COPY)
-    mock_init.reset_mock()
-
-    _ = initialize_candidate(
-        InitializationStrategy.BROADCAST_MEAN, k_strat, mock_x, mock_checkpoint
-    )
-    mock_init.assert_called_once_with(mock_x, InitializationStrategy.COPY)
-    mock_init.reset_mock()
-
-    _ = initialize_candidate(
-        InitializationStrategy.KNOWN_MATRIX, k_strat, mock_x, mock_checkpoint
-    )
-    mock_init.assert_called_once_with(mock_x, InitializationStrategy.COPY)
-
-
-@patch(f"{PKG}.initialize_low_rank_candidate")
-def test_initialize_candidate_defaults_to_copy_on_null_checkpoint(
-    mock_init: Mock,
-) -> None:
-    mock_x = np.array([[0, 2], [2, 0]])
-    mock_checkpoint = None
-    k_strat = KernelStrategy.GAUSSIAN_MODEL_SINGLE_VARIANCE
-    _ = initialize_candidate(
-        InitializationStrategy.KNOWN_MATRIX, k_strat, mock_x, mock_checkpoint
-    )
-    mock_init.assert_called_once_with(mock_x, InitializationStrategy.KNOWN_MATRIX)
-
-
-def test_initialize_candidate_throws_on_checkpoint_size_mismatch() -> None:
-    mock_x = np.eye(3)
-    mock_checkpoint = np.eye(4)
-    k_strat = KernelStrategy.GAUSSIAN_MODEL_ROWWISE_VARIANCE
-    with raises(ValueError, match="shape"):
-        _ = initialize_candidate(
-            InitializationStrategy.KNOWN_MATRIX, k_strat, mock_x, mock_checkpoint
-        )
-
-
 ## Computing max iterations
+
+
 def test_compute_max_iterations_honors_manual_count() -> None:
     manual_max = 17
     target_rank = 5
